% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOLandArea.R
\name{convertFAOLandArea}
\alias{convertFAOLandArea}
\title{Convert FAO Land Area data to complete ISO country list}
\usage{
convertFAOLandArea(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readFAOLandArea]{readFAOLandArea()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object with all ISO3c countries
(missing values set to \code{NA}).
}
\description{
Convert FAO Land Area data to complete ISO country list
}
\author{
Renato Rodrigues
}
