% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSSPextensions.R
\name{convertSSPextensions}
\alias{convertSSPextensions}
\title{Convert SSP Extensions data}
\usage{
convertSSPextensions(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from \code{\link[=readSSPextensions]{readSSPextensions()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Convert SSP Extensions data
}
\author{
Renato Rodrigues
}
