% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVDem.R
\name{convertVDem}
\alias{convertVDem}
\title{Convert V-Dem Governance Indicators}
\usage{
convertVDem(x, subtype = "all")
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readVDem]{readVDem()}}.}

\item{subtype}{passed through; \code{"all"} or a V-Dem column code.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object with all ISO3c countries
(missing values set to \code{NA}).
}
\description{
Fills missing countries with \code{NA} to ensure a complete ISO country list.
}
\author{
Renato Rodrigues
}
