% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWGIindicator.R
\name{convertWGIindicator}
\alias{convertWGIindicator}
\title{Convert World Bank Worldwide Governance Indicators}
\usage{
convertWGIindicator(x, subtype = "all")
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readWGIindicator]{readWGIindicator()}}.}

\item{subtype}{passed through; one of the WGI indicator codes or \code{"all"}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object with all ISO3c countries
(missing values set to \code{NA}).
}
\description{
Fills missing countries and returns a complete ISO country list.
}
\author{
Renato Rodrigues
}
