% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCAPMF.R
\name{downloadCAPMF}
\alias{downloadCAPMF}
\title{Download OECD Climate Actions and Policies Measurement Framework (CAPMF) data}
\usage{
downloadCAPMF()
}
\description{
Downloads the baseline CAPMF database from the OECD SDMX API, and then
generates/simulates the expanded dataset by adding stub rows for PRY, AGO, EGY, and PAK.
}
\author{
Renato Rodrigues
}
