% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpfm-package.R
\docType{package}
\name{mrpfm-package}
\alias{mrpfm}
\alias{mrpfm-package}
\title{mrpfm: Data for the Political Feasibility Module (PFM) for IAMs}
\description{
Data ingestion and preprocessing for the Political Feasibility Module (PFM), following the madrat framework (read*, calc*, convert*, tool* conventions). Provides magpie objects consumed by the pfm package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrpfm}
}

}
\author{
\strong{Maintainer}: Renato Rodrigues \email{renato.rodrigues@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5863-5514}{ORCID})

Authors:
\itemize{
  \item Elmar Kriegler \email{kriegler@pik-potsdam.de} (\href{https://orcid.org/0000-0002-3307-2647}{ORCID})
}

}
\keyword{internal}
