% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCAPMF.R
\name{readCAPMF}
\alias{readCAPMF}
\title{Read OECD Climate Actions and Policies Measurement Framework (CAPMF) data}
\usage{
readCAPMF(subtype = "all")
}
\arguments{
\item{subtype}{character, type of data: "all" or specific sector.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object with dimensions \verb{[iso3c, year, variable]}.
}
\description{
Reads the CAPMF policy stringency data from a CSV file.
}
\author{
Renato Rodrigues
}
