% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOLandArea.R
\name{readFAOLandArea}
\alias{readFAOLandArea}
\title{Read FAO Total Land Area}
\usage{
readFAOLandArea()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object with dimensions
\verb{[iso3c, year, "Land Area"]} in units of 1000 ha.
}
\description{
The FAO FAOSTAT "Inputs: Land Use - Land Cover" file contains country names
in the \code{Area} column and numeric land area values in the \code{Value} column.
Country names are converted to ISO 3166-1 alpha-3 codes using the
\code{countrycode} package.
}
\details{
Reads FAOSTAT land area data from an Excel file in the \code{FAOLandArea} source
folder. Returns a constant-in-time magpie object with land area per country
in 1000 ha.
}
\author{
Renato Rodrigues
}
