% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSSPextensions.R
\name{readSSPextensions}
\alias{readSSPextensions}
\title{Read SSP Extensions data}
\usage{
readSSPextensions()
}
\value{
MagPIE object
}
\description{
Read the input data file for the ssp extensions data
}
\author{
Renato Rodrigues
}
