% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWBCarbonPricingDashboard.R
\name{readWBCarbonPricingDashboard}
\alias{readWBCarbonPricingDashboard}
\title{Read World Bank Carbon Pricing Dashboard data}
\usage{
readWBCarbonPricingDashboard(subtype = "price")
}
\arguments{
\item{subtype}{character, type of data: price, priceAprilFirst, wbCoverage, revenue,
emissionsCovered, or the subnational variants priceSubnational / emissionsCoveredSubnational
(within_country instruments mapped to their parent country, instrument identity kept)}
}
\value{
MagPIE object with world banck data on carbon price per country
}
\description{
This data can be downloaded directly from the site.
}
\author{
Renato Rodrigues
}
