% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolImputeMedians.R
\name{toolImputeMedians}
\alias{toolImputeMedians}
\title{Impute missing values with regional or global medians}
\usage{
toolImputeMedians(
  data,
  regionMappingFiles = c("regionmapping_54.csv", "regionmappingH12.csv")
)
}
\arguments{
\item{data}{A country-level magpie object to be imputed.}

\item{regionMappingFiles}{A vector of names of the region mapping files in order of priority.
Defaults to \code{c("regionmapping_54.csv", "regionmappingH12.csv")}.}
}
\value{
An imputed magpie object of the same shape.
}
\description{
Fills NA values in a magpie object by assigning the median of the region the country belongs to.
It evaluates multiple region mapping files in order of priority. If a mapping cannot resolve a non-missing
regional median, it tries the next mapping file, eventually falling back to the global median of the year and variable.
}
\author{
Renato Rodrigues
}
