% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCAPMF.R
\name{calcCAPMF}
\alias{calcCAPMF}
\title{Calculate OECD Climate Actions and Policies Measurement Framework (CAPMF) data}
\usage{
calcCAPMF(subtype = "all", minCoverage = 0.8, includeEstimated = FALSE)
}
\arguments{
\item{subtype}{character; "all" (default) or a specific sector.}

\item{minCoverage}{numeric or logical; threshold of GDP and Population coverage below which a region is excluded. Defaults to 0.8. Set to FALSE or NULL to disable.}

\item{includeEstimated}{logical; if TRUE, estimates the composite index scores for newly onboarded developing countries.}
}
\value{
A list with:
\describe{
\item{x}{\code{\link[magclass:magclass-package]{magpie}} object \verb{[iso3c, year, variable]}}
\item{weight}{SSP2 population weights for spatial aggregation}
\item{unit}{\code{"index (0-10)"}}
\item{description}{character description}
}
}
\description{
Wraps \code{\link[madrat:readSource]{madrat::readSource}} for CAPMF data and returns sectoral stringency indicators
as a single magpie object for use as a \code{calcOutput}.
}
\author{
Renato Rodrigues
}
