% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOLandArea.R
\name{calcFAOLandArea}
\alias{calcFAOLandArea}
\title{Calculate FAO Total Land Area}
\usage{
calcFAOLandArea()
}
\value{
A list with elements:
\describe{
\item{x}{\code{\link[magclass:magclass]{magpie}} object \verb{[iso3c, year, "Land Area"]}}
\item{weight}{\code{NULL}}
\item{unit}{\code{"1000 ha"}}
\item{description}{character description}
}
}
\description{
Land area is constant in time. The function returns the most recent available
year's data. Missing countries are set to \code{NA}.
}
\details{
Wraps \code{\link[madrat:readSource]{readSource}} for FAO FAOSTAT land area data and returns it as a
\code{calcOutput} compatible magpie object.
}
\author{
Renato Rodrigues
}
