% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSPextensions.R
\name{calcSSPextensions}
\alias{calcSSPextensions}
\title{Calculate SSP Extensions data}
\usage{
calcSSPextensions(subtype = "all")
}
\arguments{
\item{subtype}{subtype "all" or "drivers_SSP2"}
}
\value{
A list with a \code{\link[magclass:magclass]{magpie}} object, weight, unit and description.
}
\description{
Calculate SSP Extensions data
}
\author{
Renato Rodrigues
}
