% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCAPMF.R
\name{convertCAPMF}
\alias{convertCAPMF}
\title{Convert OECD Climate Actions and Policies Measurement Framework (CAPMF) data}
\usage{
convertCAPMF(x, subtype = "all")
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from \code{\link[=readCAPMF]{readCAPMF()}}.}

\item{subtype}{passed through; "all" or specific sector.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object with all ISO3c countries (missing values set to \code{NA}).
}
\description{
Fills missing countries with \code{NA} to ensure a complete ISO country list.
}
\author{
Renato Rodrigues
}
