% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolTimeInterpolation.R
\name{toolTimeInterpolation}
\alias{toolTimeInterpolation}
\title{toolTimeInterpolation}
\usage{
toolTimeInterpolation(x, interpolatedYears)
}
\arguments{
\item{x}{magpie object}

\item{interpolatedYears}{years to be interpolated}

\item{backwardExtrapolation}{backward extrapolation method}

\item{extrapolationType}{extrapolation method}

\item{integrateInterpolatedYears}{integrate interpolated years}
}
\value{
Returns the combined magpie object
}
\description{
interpolates time series data backward
}
\author{
Renato Rodrigues
}
